src/simu/requirements.txt
mxnet==0.10.0
jupyter==1.0.0
numpy=1.11.0
Pillow==3.1.2
